CollisionSystem 기초 구현 및 빌드 에러 수정 - InputComponent copy 생성자/대입 연산자 delete 처리 - BoxCollider float→int narrowing 변환 수정 - CollisionSystem 설계: Collider 직접 Register/Unregister 방식 채택 - swap_and_pop 패턴으로 O(1) Unregister 구현

+++ Engine/Math/Vector3.cpp
@@ -2,7 +2,7 @@
 
 namespace Engine
 {
-	Vector3::Vector3(int x, int y, int z)
+	Vector3::Vector3(float x, float y, float z)
 		: x(x), y(y), z(z)
 	{
 	}
